% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_fp.r
\name{gg_fp}
\alias{gg_fp}
\title{gg_fp}
\usage{
gg_fp(
  data,
  x_axis,
  point,
  low,
  up,
  group_var,
  facet_var,
  zero_line = 0,
  point_space = 0.6,
  font_size = 4,
  facet_color,
  group_color,
  point_shape,
  label.x = "x axis",
  label.y = "y axis",
  point.digit = 0.01,
  parsed = FALSE
)
}
\arguments{
\item{data}{data used in \code{gg_fp()}. A \code{data.frame}.}

\item{x_axis}{variable showed in \code{x} axis. A column in \code{data} using Non-standard evaluation.}

\item{point}{point value showed in forest polt. A column in \code{data} using Non-standard evaluation.}

\item{low}{the lower 95\% CI for point value. A column in \code{data} using Non-standard evaluation.}

\item{up}{the upper 95\% CI for point value. A column in \code{data} using Non-standard evaluation.}

\item{group_var}{variable used for identify the group. A column in \code{data} using Non-standard evaluation.}

\item{facet_var}{variable used for identify the facet. A column in \code{data} using Non-standard evaluation.}

\item{zero_line}{A dotted line in forest plot. The default value is 0. A \code{1} length numeric vector.}

\item{point_space}{The spacing between points in the same group. A \code{1} length numeric vector.}

\item{font_size}{Font size in forest plot. A \code{1} length numeric vector.}

\item{facet_color}{facet color in forest plot. A character vector. \code{facet_color} length must same as \code{unique(facet_var)}.}

\item{group_color}{group color in forest plot. A character vector. \code{group_color} length must same as \code{unique(group_var)}.}

\item{point_shape}{point shape in forest plot. An integer vector. \code{point_shape} length must same as \code{unique(group_var)}.}

\item{label.x}{x label in in forest plot. The default is 'x axis'. A \code{1} length character vector.}

\item{label.y}{y label in in forest plot. The default is 'y axis'. A \code{1} length character vector.}

\item{point.digit}{The decimal number in forest plot. The default is \code{2} decimal digits \code{(point.digit = 0.01)}.
A \code{1} length character vector. You can set it to 0.1 to display 1 decimal digit. Set 0.001 to display 3 decimal digits, etc.}

\item{parsed}{A logical vector. The default is \code{FALSE}. \code{parse} indicates whether the strip in facet needs to be parsed.}
}
\value{
a plot with the class of \code{gg} and \code{ggplot}.
}
\description{
make a forest plot wiht ggplot2.
}
